allow parsers to be applied to additional filetypes, skip wezterm installation if already installed

+++ .config/yadm/lib/wezterm.sh
@@ -1,10 +1,23 @@
 #!/usr/bin/env zsh
 
-set -euo pipefail
-
 # https://wezfurlong.org/wezterm/install/linux.html
-curl -fsSL https://apt.fury.io/wez/gpg.key | sudo gpg --yes --dearmor -o /usr/share/keyrings/wezterm-fury.gpg
-echo 'deb [signed-by=/usr/share/keyrings/wezterm-fury.gpg] https://apt.fury.io/wez/ * *' | sudo tee /etc/apt/sources.list.d/wezterm.list
+wezterm_install() {
+    if dpkg -s "wezterm" > /dev/null; then
+        return
+    fi
 
-sudo apt update --yes
-sudo apt install --yes wezterm
+    local gpg_file="/usr/share/keyrings/wezterm-fury.gpg"
+    if [[ ! -f $gpg_file ]]; then
+        curl -fsSL https://apt.fury.io/wez/gpg.key | sudo gpg --yes --dearmor -o "${gpg_file}"
+    fi
+
+    local sources_file="/etc/apt/sources.list.d/wezterm.list"
+    if [[ ! -f $sources_file ]]; then
+        echo "deb [signed-by=${gpg_file}] https://apt.fury.io/wez/ * *" | sudo tee "${sources_file}"
+    fi
+
+    sudo apt update --yes
+    sudo apt install --yes wezterm
+}
+
+wezterm_install
